chore: use i8 where can

+++ src/state.rs
@@ -13,13 +13,13 @@ pub enum ProcessingMode {
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct MusicalSettings {
     /// Musical key (0-23, see keys module for mapping)
-    pub key: i32,
+    pub key: i8,
     /// Specific note (0 = auto mode, 1-9 = specific note in scale)
-    pub note: i32,
+    pub note: i8,
     /// Octave setting
-    pub octave: i32,
+    pub octave: i8,
     /// Formant shift mode (0 = none, 1 = lower, 2 = higher)
-    pub formant: i32,
+    pub formant: i8,
     /// Processing mode for vocal effects
     pub mode: ProcessingMode,
 }
